importer/requirements.txt
httpx>=0.27,<0.28
python-dotenv>=1.0,<2
pydantic>=2.8,<3
rich>=13.9,<14
typer>=0.12,<0.13
python-slugify>=8.0,<9
click>=8.1.7,<8.2

